Allow not starting Spotify. Improved play/pause toggle to explicit play and pause.
In order to use this library as part of a script which ensures that Spotify is paused, I have added the capability to specify that Spotify should not be started if it's not already open. Play/pause was previously a toggle behaviour, and explicitly ensuring playing had stopped was not possible. This functionality has been changed so that the post-condition of Pause is always stopped, and the post-condition of Play is playing (as long as Spotify has something to play in the context)